+++ pappl/system-accessors.c
@@ -1069,7 +1069,7 @@ papplSystemGetVersions(
     pthread_rwlock_rdlock(&system->rwlock);
 
     if (max_versions > (int)system->num_versions)
-      memcpy(versions, system->versions, system->num_versions * sizeof(pappl_version_t));
+      memcpy(versions, system->versions, (size_t)system->num_versions * sizeof(pappl_version_t));
     else
       memcpy(versions, system->versions, (size_t)max_versions * sizeof(pappl_version_t));
 
@@ -1175,7 +1175,7 @@ papplSystemIteratePrinters(
     pappl_printer_cb_t cb,		// I - Callback function
     void               *data)		// I - Callback data
 {
-  size_t		i,		// Looping var
+  cups_len_t		i,		// Looping var
 			count;		// Number of printers
 
 
@@ -1203,7 +1203,7 @@ papplSystemMatchDriver(
     pappl_system_t *system,		// I - System
     const char     *device_id)		// I - IEEE-1284 device ID string
 {
-  size_t	i;			// Looping var
+  cups_len_t	i;			// Looping var
   pappl_pr_driver_t *driver;		// Current driver
   const char	*drvstart,		// Start of key/value pair
 		*drvend,		// End of key/value pair
@@ -2053,7 +2053,7 @@ papplSystemSetPrinterDrivers(
   {
     pthread_rwlock_wrlock(&system->rwlock);
 
-    system->num_drivers   = (size_t)num_drivers;
+    system->num_drivers   = (cups_len_t)num_drivers;
     system->drivers       = drivers;
     system->autoadd_cb    = autoadd_cb;
     system->create_cb     = create_cb;
@@ -2158,9 +2158,9 @@ papplSystemSetVersions(
     if (num_versions > (int)(sizeof(system->versions) / sizeof(system->versions[0])))
       system->num_versions = sizeof(system->versions) / sizeof(system->versions[0]);
     else
-      system->num_versions = (size_t)num_versions;
+      system->num_versions = (cups_len_t)num_versions;
 
-    memcpy(system->versions, versions, system->num_versions * sizeof(pappl_version_t));
+    memcpy(system->versions, versions, (size_t)system->num_versions * sizeof(pappl_version_t));
 
     pthread_rwlock_unlock(&system->rwlock);
   }